modificacion campo modelo consumo materiales

+++ api_prenar/models/consumo_material.py
@@ -11,7 +11,7 @@ class ConsumoMaterial(models.Model):
     quantity_produced=models.IntegerField(null=True, blank=True)
     base_quantity_used=models.FloatField(null=True, blank=True)
     quantity_mortar_used=models.FloatField(null=True, blank=True)
-    total=models.IntegerField(default=0)
+    total=models.FloatField(default=0.0)
     email_user=models.EmailField()
     registration_date=models.DateField(auto_now_add=True)
     unit_X_base_package=models.FloatField(null=True, blank=True)
